LLM_CHAT_SIMILAR/requirements.txt
langchain_community
streamlit
openai
tiktoken
python-dotenv
faiss-cpu